Moved metadata-related validation into directory
